Move initial parameters to mlcubeutils
