+++ examples/quickstart_mlcube/server.py
@@ -1,19 +1,11 @@
-import os
 import flwr as fl
 from flwr.server.strategy import FedAvg
-import tensorflow as tf
 
-
-def initial_parameters():
-    """Return initial checkpoint parameters"""
-    package_directory = os.path.dirname(os.path.abspath(__file__))
-    filepath = os.path.join(package_directory, "mlcube/workspace/initial_checkpoint")
-    model = tf.keras.models.load_model(filepath)
-    return model.get_weights()
+import mlcube_utils as mlcube
 
 
 def main():
-    strategy = FedAvg(initial_parameters=initial_parameters())
+    strategy = FedAvg(initial_parameters=mlcube.initial_parameters())
     fl.server.start_server(
         "0.0.0.0:8080", strategy=strategy, config={"num_rounds": 3},
     )
